requirements.txt
python-json-logger==0.1.11
RPi.GPIO==0.6.5